requirements.txt
flask==2.0.1
werkzeug==2.0.3
flask-cors==4.0.0
google-generativeai==0.3.2
python-dotenv==0.19.0 